PW設定をコンポーネントに分ける
