+++ components/Setting/SettingPW.tsx
@@ -1,8 +1,14 @@
-import {Button, FormControl, FormLabel, Heading, Input} from '@chakra-ui/react';
+import {
+  Button,
+  FormControl,
+  FormLabel,
+  Heading,
+  Input,
+  Center,
+  Box,
+} from '@chakra-ui/react';
 import {useForm} from 'react-hook-form';
 
-// import {useRecoilState} from 'recoil';
-// import {UserState} from '../../utils/atom';
 type SettingInputs = {
   now_password: string;
   new_password: string;
@@ -19,31 +25,33 @@ const SettingPW = () => {
   // const toast = useToast();
 
   return (
-    <div>
-      <Heading textAlign="center" mb="1rem" size="lg">
-        パスワードパスワード変更
-      </Heading>
-      <form>
-        <FormControl mb="1rem" isDisabled>
-          <FormLabel>認証のため現在のパスワードを入力して下さい</FormLabel>
-          <Input id="" type="" placeholder="現在のパスワードを入力" />
-        </FormControl>
+    <Center>
+      <Box mt="2rem" w={{base: '97%', sm: '400px', md: '500px'}}>
+        <Heading textAlign="center" mb="1rem" size="lg">
+          パスワードパスワード変更
+        </Heading>
+        <form>
+          <FormControl mb="1rem" isDisabled>
+            <FormLabel>認証のため現在のパスワードを入力して下さい</FormLabel>
+            <Input id="" type="" placeholder="現在のパスワードを入力" />
+          </FormControl>
 
-        <FormControl mb="1rem" isDisabled>
-          <FormLabel>新しいパスワードを設定</FormLabel>
-          <Input id="" type="" placeholder="新しいパスワードを入力" />
-        </FormControl>
-        <Button
-          disabled={true}
-          isLoading={isSubmitting}
-          type="submit"
-          size="lg"
-          marginTop="2rem"
-        >
-          保存
-        </Button>
-      </form>
-    </div>
+          <FormControl mb="1rem" isDisabled>
+            <FormLabel>新しいパスワードを設定</FormLabel>
+            <Input id="" type="" placeholder="新しいパスワードを入力" />
+          </FormControl>
+          <Button
+            disabled={true}
+            isLoading={isSubmitting}
+            type="submit"
+            size="lg"
+            marginTop="2rem"
+          >
+            保存
+          </Button>
+        </form>
+      </Box>
+    </Center>
   );
 };
 export default SettingPW;
